several corrections as requested

+++ libs/pilight/json.c
@@ -1395,7 +1395,7 @@ bool json_check(const JsonNode *node, char errmsg[256])
 int json_find_numberUl(JsonNode *object, const char *name, unsigned long long *out) {
 	JsonNode *node = json_find_member(object, name);
 	if (node && node->tag == JSON_NUMBER) {
-		*out = node->number_;
+		*out = (unsigned long long)node->number_;
 		return 0;
 	}
 	return 1;
